% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterStats_HMLET.R
\name{ClusterStats_HMLET}
\alias{ClusterStats_HMLET}
\title{Generate unique random permutations
This is a sample description}
\usage{
ClusterStats_HMLET(data, paired = T, detailed = F, threshold_t = NA)
}
\description{
Generate unique random permutations
This is a sample description
}
