% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeSamplingDuration_HMLET.R
\name{ComputeSamplingDuration_HMLET}
\alias{ComputeSamplingDuration_HMLET}
\title{Compute the sampling time duration \cr}
\usage{
ComputeSamplingDuration_HMLET(data, groupingVars, timePoint = "timeStamp")
}
\arguments{
\item{data}{long format dataframe containing temporal data.}

\item{groupingVars}{string for the column(s) name for grouping variable(s) to uniquely identify datapoints within each trial}

\item{timePoint}{string for the column name that represents time intervals, for example "timeStamp."}
}
\description{
This function computs the sampling duration for each trial. Grouping variable
has to be defined to be unique for all datapoints in one trial.
}
