% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeSamplingInterval_HMLET.R
\name{ComputeSamplingInterval_HMLET}
\alias{ComputeSamplingInterval_HMLET}
\title{Compute the sampling time interval}
\usage{
ComputeSamplingInterval_HMLET(
  data,
  ID = "ID",
  trial,
  timePoint,
  appendFlag = T,
  userCallFlag = T
)
}
\arguments{
\item{data}{long format dataframe containing temporal data.}

\item{ID}{string for column name that represents IDs within data frame, defaults to "ID".}

\item{trial}{string for column name that represents trials within data frame. Can be numerical or categorical.}

\item{timePoint}{string for column name that represents time stamps of each sample.}

\item{appendFlag}{optional flag specifies whether the interval(s) being appended to the data or returned separately, defaults to TRUE.}

\item{userCallFlag}{optional flag to print out data when users are calling this function, defaults to TRUE.}
}
\description{
This function estimates the sampling interval for each trial based on the time
differences between each time point and uses mode to find the interval.
Note that your dataset has to have a unique consistent time interval between all samples.
In case, several time intervals is estimated for your dataset it use "Preprocess_FixSampling_HMLET"
function and pass the highest sampling interval to fix the inconsistency.
}
