% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeSubjectLevelPerm_HMLET.R
\name{ComputeSubjectLevelPerm_HMLET}
\alias{ComputeSubjectLevelPerm_HMLET}
\title{ComputeSubjectLevelPerm_HMLET
Provides unique permututation indices within each subject.}
\usage{
ComputeSubjectLevelPerm_HMLET(labels, n = 1)
}
\arguments{
\item{labels}{labels from the dataframe}

\item{n}{optional integer for number of permutations for resampling data}
}
\value{
labels of unique permutation indices to use for subject level permutation test
}
\description{
ComputeSubjectLevelPerm_HMLET
Provides unique permututation indices within each subject.
}
