% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeTValues_HMLET.R
\name{ComputeTValues_HMLET}
\alias{ComputeTValues_HMLET}
\title{ComputeTValues_HMLET}
\usage{
ComputeTValues_HMLET(respTime, paired = TRUE)
}
\arguments{
\item{respTime}{dataframe for responses in temporal order.}

\item{paired}{optional boolean to control T-tests. True is for statistical comparisons in paired mode,
False for a two-sample T-Test; defaults to True.}
}
\value{
t values in the format of a data frame.
}
\description{
ComputeTValues_HMLET
}
