% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportDataForGUI_HMLET.R
\name{ExportDataForGUI_HMLET}
\alias{ExportDataForGUI_HMLET}
\title{Export to CSV file for MATLAB GUI}
\usage{
ExportDataForGUI_HMLET(
  data,
  ID = "ID",
  trial = "trial",
  timepoint = "timepoint",
  timeMax = NULL,
  dataPointDuration,
  response,
  fixation,
  condition,
  testName = NULL,
  gazeX,
  gazeY,
  gazeXRelative = NULL,
  gazeYRelative = NULL,
  miscVars = NULL,
  fileName = "HMLET_DataforMATLAB.csv",
  path = getwd()
)
}
\description{
Export to CSV file for MATLAB GUI
}
