% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateRandomData_HMLET.R
\name{GenerateRandomData_HMLET}
\alias{GenerateRandomData_HMLET}
\title{Generate unique random permutations
To be completed}
\usage{
GenerateRandomData_HMLET(
  tMax = 20,
  effectOffset = 5,
  trialNum = 40,
  subjNum = 20,
  effectSize = 0.2
)
}
\description{
Generate unique random permutations
To be completed
}
