% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermutationTest_HMLET.R
\name{PermutationTest_HMLET}
\alias{PermutationTest_HMLET}
\title{Permutation Tests General Routine}
\usage{
PermutationTest_HMLET(
  data,
  samples = 2000,
  paired = T,
  permuteTrialsWithinSubject = F,
  threshold_t = NA
)
}
\arguments{
\item{data}{data frame with temporal data for permutation tests routine, output from PermutationTestDataPrep_HMLET.R}

\item{samples}{optional integer for number of samples, defaults to 2000}

\item{paired}{optional boolean XX, defaults to True # ASK QUESTION}

\item{permuteTrialsWithinSubject}{optional boolean to run permutation test pipeline in subject level or trial level, defaults to False to run trial level}

\item{threshold_t}{optional value XX, defaults to NA # ASK QUESTION}
}
\value{
a list of permutation tests in which distribution of desired statistic under the null hypothesis is estimated in a large number of permutations of the original data
}
\description{
Permutation Tests General Routine
}
