% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermutationTestDataPrep_HMLET.R
\name{PermutationTestDataPrep_HMLET}
\alias{PermutationTestDataPrep_HMLET}
\title{Return a data frame compatible for permutation tests routine.}
\usage{
PermutationTestDataPrep_HMLET(
  data,
  ID = "ID",
  trial = NULL,
  timePoint,
  condition,
  gazeMeasure,
  conditionLevels = NULL,
  targetAOI = NULL,
  testName = NULL,
  timeBinName = NULL
)
}
\arguments{
\item{data}{data frame with temporal data for permutation tests routine.}

\item{ID}{optional string for column name that represents IDs within data frame, defaults to "ID".}

\item{trial}{string for column name that represents trials within data frame.}

\item{timePoint}{string for column name that includes time point or index values in numeric form.}

\item{condition}{string for column name that specifies a condition within the data frame.}

\item{gazeMeasure}{string for column name which includes numerical value of the gaze measure at each time point such as gazeproportion.}

\item{conditionLevels}{optional array of string column names to denote all the values of conditions, defaults to NULL.}

\item{targetAOI}{optional string specify the name of the target AOI which should be set as 1 in the gazeMeasure variable, defaults to NULL.}

\item{testName}{optional string for name of data -- used as condition name or test names to compare permutation test
results between different tests/conditions later, defaults to NULL.}

\item{timeBinName}{optional string for the column name that includes time bin names, defaults to NULL.}
}
\value{
data frame compatible for permutation tests and other HMLET modules
}
\description{
Return a data frame compatible for permutation tests routine.
}
