% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermutationTestSingleParticipant_HMLET.R
\name{PermutationTestSingleParticipant_HMLET}
\alias{PermutationTestSingleParticipant_HMLET}
\title{Conduct trial level permutation test for single participant.}
\usage{
PermutationTestSingleParticipant_HMLET(data, samples = 2000, threshold_t = NA)
}
\arguments{
\item{data}{data frame with temporal data for permutation tests routine, output from PermutationTestDataPrep_HMLET.R.}

\item{samples}{optional integer for number of samples, defaults to 2000.}

\item{threshold_t}{optional probability threshold for statistical comparison, defaults to NA and will be computed based on number of trials when "between trials
permutation" is called or will be computed based on number of subjects when "between subjects permutation" is called.
Alpha = 0.025.}
}
\value{
a list of permutation tests in which distribution of desired statistic under the null hypothesis is estimated in a large number of permutations of the original data
}
\description{
Conduct trial level permutation test for single participant.
}
