% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTemporalGazeTrends_HMLET.R
\name{PlotTemporalGazeTrends_HMLET}
\alias{PlotTemporalGazeTrends_HMLET}
\title{Plot temporal gaze trends.}
\usage{
PlotTemporalGazeTrends_HMLET(
  resultList,
  showDataPointNumbers = T,
  showOverallMean = T,
  gazePropRibbonAlpha = 0.1,
  clusterFillColor = "#CC9933",
  clusterFillAlpha = 0.5,
  pointSize = 1,
  pointAlpha = 0.7,
  pointFatten = 3,
  testNameOrder = NULL,
  conditionOrder = NULL,
  onlySignificantClusters = T,
  clusterData = NULL,
  yLabel = "Gaze Proportion",
  dataAxisLabel = "Data Points (\%)",
  xLabel = "Time (ms)",
  lineWidthGazeProp = 1,
  lineWidthDataPointNumbers = 1,
  lineWidthOverallMean = 1,
  lineTypeOverallMean = "dashed",
  alphaOverallMean = 0.5
)
}
\arguments{
\item{resultList}{dataframe of the data that is already prepared by PrepareMLETData_HMLET
or a list that is the result of PermutationTest_HMLET.}

\item{showDataPointNumbers}{optional boolean to display data points on the axes, defaults to True.}

\item{showOverallMean}{optional boolean to display a horizontal line on the y = average across all time points, defaults to True.}

\item{gazePropRibbonAlpha}{optional float to change opacity of gaze ribbon, defaults to 0.1.}

\item{clusterFillColor}{optional string to change fill color of clustered data points, defaults to "#CC9933".}

\item{clusterFillAlpha}{optional float to change opacity of clustered data points, defaults to 0.5.}

\item{pointSize}{optional integer to change size of points on plot, defaults to 1.}

\item{pointAlpha}{optional float to change opacity of points on plot, defaults to 0.7.}

\item{pointFatten}{optional integer to change size of lines on plot, defaults to 3.}

\item{testNameOrder}{specify order of test name labels, defaults to NULL.}

\item{conditionOrder}{specify order of condition levels, defaults to NULL.}

\item{onlySignificantClusters}{optional Boolean to plot only the significant clusters, defaults to True.}

\item{clusterData}{data containing all clusters generated from ClusterStats_HMLET to display on plot, defaults to NULL.}

\item{yLabel}{= "Gaze Proportion" specify label for the y axis.}

\item{dataAxisLabel}{= "Data Points (%)" specify label for the second y axis on the right.}

\item{xLabel}{= "Time (ms)" specify label for the x axis.}

\item{lineWidthGazeProp}{= 1 specify line width for gaze proportion}

\item{lineWidthDataPointNumbers}{= 1 specify line width for data point numbers}

\item{lineWidthOverallMean}{= 1 specify line width for overall mean}

\item{lineTypeOverallMean}{= "dashed" specify line type for overall mean}

\item{alphaOverallMean}{= 0.5 specify alpha for overall mean}
}
\value{
a plot handle that visualizes the data from PrepareMLETData_HMLET or the list from PermuationTest_HMLET.
}
\description{
Plot temporal gaze trends.
}
