% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess_AOIExtraction_HMLET.R
\name{Preprocess_AOIExtraction_HMLET}
\alias{Preprocess_AOIExtraction_HMLET}
\title{AOI Extraction}
\usage{
Preprocess_AOIExtraction_HMLET(
  data,
  GazeX,
  GazeY,
  AOI_Names,
  AOIs_Center_X,
  AOIs_Center_Y,
  AOIs_Widths,
  AOIs_Heights,
  ContentLabel = "Content",
  EllipticAOI = FALSE
)
}
\arguments{
\item{data}{long format dataframe containing temporal data.}

\item{GazeX}{string for column name that includes X coordinate of gazepoints}

\item{GazeY}{string for column name that includes Y coordinate of gazepoints}

\item{AOI_Names}{List of Strings for AOIs names}

\item{AOIs_Center_X}{List of numerical values for the horizontal coordinate of the center of all AOIs or list of strings indicating the column names in the input data that have these values}

\item{AOIs_Center_Y}{List of numerical values for the vertical coordinate of the center of all AOIs or list of strings indicating the column names in the input data that have these values}

\item{ContentLabel}{Label for gaze points that are outside all of the AOIs, defaults to "Content"}

\item{EllipticAOI}{Binary value to identify whether elliptic AOIs should be used instead of rectangular ones, defaults to FALSE}

\item{AOIs_Radius_X}{List of numerical values for width size/horizontal diameter of all AOIs or list of strings indicating the column names in the input data that have these values}

\item{AOIs_Radius_Y}{List of numerical values for height size/vertical diameter of all AOIs or list of strings indicating the column names in the input data that have these values.}
}
\description{
AOIs can be computed in rectangular or elliptic forms.
This function uses gaze coordinates to categorize them into specified AOIs
}
\details{
The output is the same dataframe as input with one new column:
AOIs: Includes the name of the AOI at each time point.
}
