% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess_CheckData_HMLET.R
\name{Preprocess_CheckData_HMLET}
\alias{Preprocess_CheckData_HMLET}
\title{Check Data for Critical Issues}
\usage{
Preprocess_CheckData_HMLET(
  data,
  ID,
  trial,
  timePoint,
  GazeX,
  GazeY,
  missRateCap = 50
)
}
\arguments{
\item{data}{long format dataframe containing temporal data.}

\item{ID}{string for column name that represents IDs within data frame, defaults to "ID".}

\item{trial}{string for column name that identify unique trials within data frame. Can be numerical or categorical.}

\item{timePoint}{string for column name that represents time stamps of each sample.}

\item{GazeX}{string for column name that includes X coordinate of gazepoints}

\item{GazeY}{string for column name that includes Y coordinate of gazepoints}

\item{missRateCap}{Specify the percentage of missing gazepoints cap to be counted as high Miss Rate}
}
\description{
Raw eye-tracking data comes with time stamps of each sample and Gaze Coordinates. These time stamps
has to have consistent and unique time interval across all datapoints.
Gaze Coordinates may have missing values due to blinks, head movement, etc.
This function reports issues about inconsistent sampling intervals and missing Gaze coordinates.
}
