% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreprocessRawData_HMLET.R
\name{PreprocessRawData_HMLET}
\alias{PreprocessRawData_HMLET}
\title{Pre-Process data to extract AOIs. \cr
Creates a csv file that formats all data into a table. \cr
Should be Completed later}
\usage{
PreprocessRawData_HMLET(
  data,
  ID = "ID",
  trial = "trial",
  timePoint = "timePoint",
  GazeX = "GazeX_Relative",
  GazeY = "GazeY_Relative",
  AOINames = NULL,
  fileName = "ETDataforMATLAB.csv",
  path = getwd()
)
}
\arguments{
\item{data}{dataframe containing temporal data.}

\item{ID}{optional string for column name that represents IDs within data frame, defaults to "ID."}

\item{trial}{string for column name that represents trials within data frame, for example "TrialNum."}

\item{timePoint}{string for column name that represents time intervals, for example "timeStamp."}

\item{GazeX}{# TODO: ASK}

\item{GazeY}{# TODO: ASK}

\item{fileName}{optional string for output csv file, defaults to "ETDataforMATLAB.csv."}

\item{path}{specifies working directory of project.}
}
\description{
Pre-Process data to extract AOIs. \cr
Creates a csv file that formats all data into a table. \cr
Should be Completed later
}
