% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelativeViewingProportionLabels_HMLET.R
\name{RelativeViewingProportionLabels_HMLET}
\alias{RelativeViewingProportionLabels_HMLET}
\title{Label trials whether participant was disproportionately or equally looking toward items}
\usage{
RelativeViewingProportionLabels_HMLET(
  data,
  ID = "ID",
  trial,
  AOIName,
  comparingAOI = c("Target", "Lure"),
  targetAOI = "Target",
  dispropotionCriterion = 10
)
}
\arguments{
\item{data}{data frame with temporal data for permutation tests routine.}

\item{ID}{optional string for column name that represents IDs within data frame, defaults to "ID".}

\item{trial}{string for column name that represents trials within data frame.}

\item{AOIName}{string for column name which includes name of the AOI that the subject is looking at each timepoint.}

\item{comparingAOI}{names of the two AOIs that are their looking proportion should be compared.}

\item{targetAOI}{names of the two AOIs that are their looking proportion should be compared.}

\item{dispropotionCriterion}{optional string for column name to focus on specific AOI, defaults to NULL.}

\item{testName}{Name of this data -- can be used as condition name or test names to compare permutation test results between different tests/conditions later}
}
\value{
data frame as original data with an extracolumn with labels
}
\description{
Label trials whether participant was disproportionately or equally looking toward items
}
