% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubjectLevelPermutationTestWithin_HMLET.R
\name{SubjectLevelPermutationTestWithin_HMLET}
\alias{SubjectLevelPermutationTestWithin_HMLET}
\title{SubjectLevelPermutationTestWithin_HMLET}
\usage{
SubjectLevelPermutationTestWithin_HMLET(
  data,
  samples = 2000,
  paired = T,
  threshold_t = NA
)
}
\arguments{
\item{data}{data frame with temporal data for permutation tests routine, output from PermutationTestDataPrep_HMLET.R.}

\item{samples}{optional number of resampling data, how many permutations needed.}

\item{paired}{optional boolean to control T-tests. True is for statistical comparisons in paired mode,
False for a two-sample T-Test; defaults to True.}

\item{threshold_t}{optional probability threshold for statistical comparison computed based on number of subjects.
alpha = 0.025.}
}
\value{
returns t value distribution presented in a data frame.
}
\description{
SubjectLevelPermutationTestWithin_HMLET
}
