% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrialLevelPermutationTestWithin_HMLET.R
\name{TrialLevelPermutationTestWithin_HMLET}
\alias{TrialLevelPermutationTestWithin_HMLET}
\title{Generate unique random permutations}
\usage{
TrialLevelPermutationTestWithin_HMLET(
  data,
  samples = 2000,
  paired = T,
  threshold_t = NA
)
}
\description{
Generate unique random permutations
}
