% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniquePermutations_HMLET.R
\name{UniquePermutations_HMLET}
\alias{UniquePermutations_HMLET}
\title{Generate unique random permutations.}
\usage{
UniquePermutations_HMLET(listInput, uniqueLabels, n = 1)
}
\arguments{
\item{listInput}{Arbitrary list.}

\item{uniqueLabels}{All possible Labels.}

\item{n}{An integer number indicating the number of permutations.}
}
\description{
Generate unique random permutations.
}
