% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeSubjectLevelPerm_HMLET.R
\name{ComputeSubjectLevelPerm_HMLET}
\alias{ComputeSubjectLevelPerm_HMLET}
\title{Generate unique random permutations}
\usage{
ComputeSubjectLevelPerm_HMLET(labels, n = 1)
}
\description{
Generate unique random permutations
}
