% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTemporalGazeTrends_HMLET.R
\name{PlotTemporalGazeTrends_HMLET}
\alias{PlotTemporalGazeTrends_HMLET}
\title{Plot temporal gaze trends.}
\usage{
PlotTemporalGazeTrends_HMLET(
  resultList,
  showDataPointProp = T,
  showOverallMean = "Point",
  gazePropRibbonAlpha = 0.1,
  clusterFillColor = "#CC9933",
  clusterFillAlpha = 0.5,
  pointSize = 1,
  pointAlpha = 0.7,
  pointFatten = 3,
  testNameOrder = NULL,
  conditionOrder = NULL,
  onlySignificantClusters = T,
  clusterData = NULL,
  yLabel = "Gaze Proportion",
  dataAxisLabel = "Data Points (\%)",
  xLabel = "Time (ms)",
  lineWidthGazeProp = 1,
  lineWidthDataPointProp = 1,
  lineWidthOverallMean = 1,
  lineTypeOverallMean = "dotted",
  alphaOverallMean = 0.5,
  tickSizeOverallMean = NULL
)
}
\arguments{
\item{resultList}{dataframe of the data that is already prepared by PrepareMLETData_HMLET
or a list that is the result of PermutationTest_HMLET.}

\item{showDataPointProp}{optional boolean to display data points on the axes, defaults to True.}

\item{showOverallMean}{optional string to specify the overall mean. "Line" = to display a horizontal line on the y = average. "Point" to display overall average and range in a pointRange format. "none" to not display it, defaults to "Point".}

\item{gazePropRibbonAlpha}{optional float to change opacity of gaze ribbon, defaults to 0.1.}

\item{clusterFillColor}{optional string to change fill color of clustered data points, defaults to "#CC9933".}

\item{clusterFillAlpha}{optional float to change opacity of clustered data points, defaults to 0.5.}

\item{pointSize}{optional integer to change size of points on plot, defaults to 1.}

\item{pointAlpha}{optional float to change opacity of points on plot, defaults to 0.7.}

\item{pointFatten}{optional integer to change size of lines on plot, defaults to 3.}

\item{testNameOrder}{specify order of test name labels, defaults to NULL.}

\item{conditionOrder}{specify order of condition levels, defaults to NULL.}

\item{onlySignificantClusters}{optional Boolean to plot only the significant clusters, defaults to True.}

\item{clusterData}{data containing all clusters generated from ClusterStats_HMLET to display on plot, defaults to NULL.}

\item{yLabel}{Proportion" specify label for the y axis, defaults to "Gaze"}

\item{dataAxisLabel}{Specify label for the second y axis on the right, defaults to "Data Points \(\%\)".}

\item{xLabel}{Specify label for the x axis, defaults to "Time \(ms\)"}

\item{lineWidthGazeProp}{Specify line width for gaze proportion, defaults to 1.}

\item{lineWidthDataPointProp}{Specify line width for data point numbers, defaults to 1.}

\item{lineWidthOverallMean}{Specify line width for overall mean, defaults to 1.}

\item{lineTypeOverallMean}{Specify line type for overall mean, defaults to "dotted".}

\item{alphaOverallMean}{Specify alpha for overall mean, defaults to 0.5.}

\item{tickSizeOverallMean}{Optional to specify where to plot the overall means when showOverallMean = "Point", defaults to NULL.}
}
\value{
a plot handle that visualizes the data from PrepareMLETData_HMLET or the list from PermuationTest_HMLET.
}
\description{
Plot temporal gaze trends.
}
