% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreprocessRawData_HMLET.R
\name{PreprocessRawData_HMLET}
\alias{PreprocessRawData_HMLET}
\title{Pre-Process data to extract AOIs. \cr
Creates a csv file that formats all data into a table. \cr
Should be Completed later}
\usage{
PreprocessRawData_HMLET(
  data,
  ID = "ID",
  trial = "trial",
  timepoint = "timepoint",
  GazeX = "GazeX_Relative",
  GazeY = "GazeY_Relative",
  AOINames = NULL,
  fileName = "ETDataforMATLAB.csv",
  path = getwd()
)
}
\description{
Pre-Process data to extract AOIs. \cr
Creates a csv file that formats all data into a table. \cr
Should be Completed later
}
